GameScreen: Fix aspect ratio calculation

+++ core/src/com/ankreez/metearth/Screens/GameScreen.java
@@ -7,7 +7,8 @@ import com.badlogic.gdx.Gdx;
 
 public class GameScreen extends SimpleScreen {
 
-    private static final float ASPECT_RATIO = Gdx.graphics.getHeight() / Gdx.graphics.getWidth();
+    private static final float ASPECT_RATIO =
+            (float) Gdx.graphics.getHeight() / (float) Gdx.graphics.getWidth();
 
     private static final float WORLD_WIDTH = 128.0f;
     private static final float WORLD_HEIGHT = WORLD_WIDTH * ASPECT_RATIO;
